requirements.txt
openai

python-dotenv

langchain_openai

langchain-google-genai

langchain-ollama

langchain-huggingface

streamlit

langchain-aws